requirements.txt
Flask
os
glob
pandas
google.generativeai
numpy